Stable build and install state

+++ src/commands/initialize.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import fs from 'fs/promises';
 import fsx from 'fs-extra';
 import path from 'path';
+import { initFolderConfig } from '../config';
 
 export async function initialize() {
   if ((vscode.workspace.workspaceFolders?.length || 0) > 0) {
@@ -10,10 +11,14 @@ export async function initialize() {
       '.vscode'
     );
     await fsx.ensureDir(vscodePath);
-    await fs.writeFile(
-      path.join(vscodePath, '.ios-remote-build'),
-      '// This file just tells VSCode to activate the iOS Remote Build Extension'
-    );
+    const configFilePath = path.join(vscodePath, 'ios-remote-build.json');
+    if (!(await fsx.exists(configFilePath))) {
+      await fs.writeFile(
+        configFilePath,
+        JSON.stringify({ provisioningProfile: '', exportOptionsPlist: '', devTeamId: '', })
+      );
+    }
+    initFolderConfig();
   } else {
     vscode.window.showErrorMessage(
       'No folder active, please open a Folder/Workspace to use iOS Remote Build'
